Moving some hardware specific functions out
