+++ src/main/drivers/bus_spi_impl.h
@@ -22,18 +22,6 @@
 
 #define SPI_TIMEOUT_US  10000
 
-#if defined(STM32F4) || defined(STM32G4)
-#define MAX_SPI_PIN_SEL 2
-#elif defined(STM32F7) || defined(AT32F4)
-#define MAX_SPI_PIN_SEL 4
-#elif defined(STM32H7)
-#define MAX_SPI_PIN_SEL 5
-#elif defined(APM32F4)
-#define MAX_SPI_PIN_SEL 2
-#else
-#error Unknown MCU family
-#endif
-
 #define BUS_SPI_FREE   0x0
 
 typedef struct spiPinDef_s {
