+++ src/platform/APM32/mk/APM32F4.mk
@@ -177,6 +177,7 @@ MCU_COMMON_SRC = \
         drivers/bus_i2c_config.c \
         drivers/bus_spi_config.c \
         common/stm32/bus_spi_pinconfig.c \
+        common/stm32/bus_spi_hw.c \
         drivers/serial_escserial.c \
         drivers/serial_pinconfig.c \
         drivers/serial_uart_pinconfig.c \
@@ -202,6 +203,7 @@ MSC_SRC = \
         msc/usbd_storage_sdio.c
 
 SPEED_OPTIMISED_SRC += \
+        common/stm32/bus_spi_hw.c \
         common/stm32/system.c
 
 SIZE_OPTIMISED_SRC += \
