+++ src/platform/SIMULATOR/platform_mcu.h
@@ -21,10 +21,12 @@
 
 #pragma once
 
-#define IOCFG_OUT_PP         0
-#define IOCFG_OUT_OD         0
-#define IOCFG_AF_PP          0
-#define IOCFG_AF_OD          0
-#define IOCFG_IPD            0
-#define IOCFG_IPU            0
-#define IOCFG_IN_FLOATING    0
+#define IOCFG_OUT_PP        0
+#define IOCFG_OUT_OD        0
+#define IOCFG_AF_PP         0
+#define IOCFG_AF_OD         0
+#define IOCFG_IPD           0
+#define IOCFG_IPU           0
+#define IOCFG_IN_FLOATING   0
+
+#define SPIDEV_COUNT        0
