+++ src/platform/STM32/mk/STM32_COMMON.mk
@@ -2,7 +2,11 @@
 INCLUDE_DIRS += $(PLATFORM_DIR)/common/stm32
 
 MCU_COMMON_SRC += \
-            common/stm32/bus_spi_pinconfig.c
+            common/stm32/bus_spi_pinconfig.c \
+            common/stm32/bus_spi_hw.c \
 
 SIZE_OPTIMISED_SRC += \
             common/stm32/bus_spi_pinconfig.c
+
+SPEED_OPTIMISED_SRC += \
+            common/stm32/bus_spi_hw.c
